+++ doc/changelog.rst
@@ -60,6 +60,7 @@ Version 3.9 adds support for MongoDB 4.2. Highlights include:
   :meth:`~pymongo.collection.Collection.find_one_and_update`,
   :meth:`~pymongo.operations.UpdateOne`, and
   :meth:`~pymongo.operations.UpdateMany`.
+- New BSON utility functions :func:`~bson.encode` and :func:`~bson.decode`
 - :class:`~bson.binary.Binary` now supports any bytes-like type that implements
   the buffer protocol.
 - Resume tokens can now be accessed from a ``ChangeStream`` cursor using the